Close socket on error
