+++ src/test/java/com/transloadit/sdk/AssemblyTest.java
@@ -360,53 +360,14 @@ public class AssemblyTest extends MockHttpService {
         assertEquals(savedAssembly.json().get("ok"), "ASSEMBLY_EXECUTING");
     }
 
+
     /**
-     * Saves a multithreaded upload and verifies that the requests are sent.
-     * @throws IOException
-     * @throws LocalOperationException
-     * @throws RequestException
+     * Tests the functionality of {@link Assembly#getClientSideGeneratedAssemblyID()}.
      */
     @Test
-    public void saveMultiThreadedUpload() throws IOException, LocalOperationException, RequestException {
-        MockTusAssemblyMultiThreading assembly = new MockTusAssemblyMultiThreading(transloadit);
-        assembly.addFile(new File("LICENSE"), "file_name");
-        assembly.addFile(new File("LICENSE"), "file_name2");
-        assembly.setMaxParallelUploads(2);
-
-        String uploadSize = "1077";
-        mockServerClient.when(HttpRequest.request()
-                        .withPath("/assemblies").withMethod("POST"))
-                .respond(HttpResponse.response().withBody(getJson("resumable_assembly.json")));
-
-        AssemblyResponse response = assembly.save(true);
-
-        mockServerClient.when(
-                HttpRequest.request()
-                        .withPath("/resumable/files").withMethod("POST"), Times.exactly(1)).respond(
-                new HttpResponse()
-                        .withStatusCode(201)
-                        .withHeader("Tus-Resumable", "1.0.0").withHeader(
-                                "Location", "http://localhost:9040/resumable/files/2"));
-        mockServerClient.when(
-                HttpRequest.request()
-                        .withPath("/resumable/files").withMethod("POST"), Times.exactly(1)).respond(
-                new HttpResponse()
-                        .withStatusCode(201)
-                        .withHeader("Tus-Resumable", "1.0.0").withHeader(
-                                "Location", "http://localhost:9040/resumable/files/1"));
-
-
-        mockServerClient.when(HttpRequest.request()
-                .withPath("/resumable/files/1").withMethod("PATCH").withHeader(
-                        "Upload-Length")).respond(
-                new HttpResponse()
-                        .withStatusCode(204)
-                        .withHeader("Tus-Resumable", "1.0.0")
-                        .withHeader("Upload-Offset", uploadSize));
-
-
-        assertEquals(response.json().get("assembly_id"), "02ce6150ea2811e6a35a8d1e061a5b71");
-        assertEquals(response.json().get("ok"), "ASSEMBLY_UPLOADING");
+    public void getAssemblyId() throws LocalOperationException {
+        assembly.setAssemblyId("68fffff5474d40b8bf7a294cfce4aba5");
+        assertEquals("68fffff5474d40b8bf7a294cfce4aba5", assembly.getClientSideGeneratedAssemblyID());
     }
 
     /**
@@ -426,15 +387,6 @@ public class AssemblyTest extends MockHttpService {
             assembly.setAssemblyId(uuidLong); });
         assembly.setAssemblyId(uuid);
         assertEquals(assembly.getClientSideGeneratedAssemblyID(), uuid);
-    }
-
-    /**
-     * Tests the functionality of {@link Assembly#getClientSideGeneratedAssemblyID()}.
-     */
-    @Test
-    public void getAssemblyId() throws LocalOperationException {
-        assembly.setAssemblyId("68fffff5474d40b8bf7a294cfce4aba5");
-        assertEquals("68fffff5474d40b8bf7a294cfce4aba5", assembly.getClientSideGeneratedAssemblyID());
     }
 
     /**
